refactor(ui): move the chart tooltips into the UI library
`CursorTooltip` and `SeriesTooltipContent` had drifted into being shared
chrome without moving: between them they had twelve call sites, and only
three of those were in the folders the files lived under. Neither carries
anything from the app, so both move to `@everr/ui` as they are.

The one edit is to two doc comments in `series-tooltip.tsx`, which used
em dashes.
